SVM: move `TransactionProcessingCallback` into its own module
