+++ svm/tests/mock_bank.rs
@@ -7,7 +7,7 @@ use {
         pubkey::Pubkey,
         rent_collector::RentCollector,
     },
-    solana_svm::transaction_processor::TransactionProcessingCallback,
+    solana_svm::transaction_processing_callback::TransactionProcessingCallback,
     std::{cell::RefCell, collections::HashMap, sync::Arc},
 };
